fix(game): fix winner calculation
