+++ src/services/v1/game.ts
@@ -8,7 +8,7 @@ import { EmbeddedPlayer } from '../../types/player'
 import { Types } from 'mongoose'
 import { ApiError } from '../../types/error'
 import ITeam, { TeamData } from '../../types/team'
-import { updateGameData } from '../../utils/game-stats'
+import { calculateWinner, updateGameData } from '../../utils/game-stats'
 import AtomicTeam from '../../models/atomic-team'
 import { getInitialTeamData } from '../../utils/team-stats'
 import { IAtomicPlayer } from '../../types/atomic-stat'
@@ -126,6 +126,7 @@ export const finishGame = async (gameId: string) => {
 
     if (winner === 'one') {
         if (prevWinner === 'two') {
+            // needed when a game is restarted
             updateTeam(1, -1, teamOne)
             updateTeam(-1, 1, teamTwo)
             updateTeam(1, -1, atomicTeamOne)
@@ -142,6 +143,7 @@ export const finishGame = async (gameId: string) => {
         }
     } else {
         if (prevWinner === 'one') {
+            // needed when a game is restarted
             updateTeam(1, -1, teamTwo)
             updateTeam(-1, 1, teamOne)
             updateTeam(1, -1, atomicTeamTwo)
@@ -175,19 +177,6 @@ const updateTeam = async (wins: number, losses: number, team?: TeamData | null) 
     if (!team) return
     team.wins += wins
     team.losses += losses
-}
-
-const calculateWinner = (game: IGame): 'one' | 'two' => {
-    const scores = { teamOne: 0, teamTwo: 0 }
-    for (const point of game.points) {
-        if (point.teamOne.goalsFor === 1) {
-            scores.teamOne += 1
-        } else if (point.teamTwo.goalsFor === 1) {
-            scores.teamTwo += 1
-        }
-    }
-
-    return scores.teamOne >= scores.teamTwo ? 'one' : 'two'
 }
 
 export const getGameById = async (gameId: string): Promise<IGame> => {
